+++ R/predict_glm.R
@@ -5,11 +5,15 @@
 #' @param var_correct A logical indicating whether to return the corrected prediction
 #'   variances when predicting via models fit using \code{ssn_glm}. The default is
 #'   \code{TRUE}.
+#' @param dispersion The dispersion of assumed when computing the prediction standard errors
+#'   for \code{ssn_glm()} model objects when \code{family}
+#'   is \code{"nbinomial"}, \code{"beta"}, \code{"Gamma"}, or \code{"inverse.gaussian"}.
+#'   If omitted, the model object dispersion parameter is used.
 #' @rdname predict.SSN2
 #' @method predict ssn_glm
 #' @export
-predict.ssn_glm <- function(object, newdata, type = c("link", "response"), se.fit = FALSE, interval = c("none", "confidence", "prediction"),
-                            newdata_size, level = 0.95, var_correct = TRUE, local, ...) {
+predict.ssn_glm <- function(object, newdata, type = c("link", "response", "terms"), se.fit = FALSE, interval = c("none", "confidence", "prediction"),
+                            level = 0.95, dispersion = NULL, terms = NULL, local, var_correct = TRUE, newdata_size, na.action = na.fail, ...) {
   # match type argument so the two display
   type <- match.arg(type)
 
@@ -22,6 +26,14 @@ predict.ssn_glm <- function(object, newdata, type = c("link", "response"), se.fi
 
   # deal with newdata_size
   if (missing(newdata_size)) newdata_size <- NULL
+
+  # handle dispersion argument if provided
+  if (!is.null(dispersion)) {
+    if (object$family %in% c("binomial", "poisson") && dispersion != 1) {
+      stop("dispersion is fixed at one for binomial and poisson families.", call. = FALSE)
+    }
+    object$coefficients$params_object$dispersion[1] <- dispersion
+  }
 
   # handle local for now
   if (missing(local)) local <- NULL
@@ -111,6 +123,13 @@ predict.ssn_glm <- function(object, newdata, type = c("link", "response"), se.fi
   newdata_model <- newdata_model[, keep_cols, drop = FALSE]
   attr(newdata_model, "assign") <- attr_assign[keep_cols]
   attr(newdata_model, "contrasts") <- attr_contrasts
+
+  # call terms if needed
+  if (type == "terms") {
+    # glm supports standard errors for terms objects but not intervals (no interval argument)
+    # scale df not used for glms
+    return(predict_terms(object, newdata_model, se.fit, scale = NULL, df = Inf, interval, level, add_newdata_rows, terms, ...))
+  }
 
   # storing newdata as a list
   newdata_rows_list <- split(newdata, seq_len(NROW(newdata)))